improvements in MyItemScreen

+++ app/src/main/java/com/mego/clothy/ui/HomeViewModel.kt
@@ -16,6 +16,7 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.mapLatest
 import kotlinx.coroutines.flow.stateIn
@@ -25,6 +26,7 @@ import java.io.FileOutputStream
 import java.io.InputStream
 import java.text.SimpleDateFormat
 import java.util.Locale
+import java.util.stream.Collectors
 import javax.inject.Inject
 
 const val SELECTED_CATEGORY_KEY = "selected_category"
@@ -45,7 +47,9 @@ class HomeViewModel @Inject constructor (
 
     private val actionModeSelectedItemsList = mutableListOf<Item>()
 
-    var selectedItemToShow : Item? = null
+    private val _selectedItemToShow = MutableStateFlow<Item?>(null)
+    val selectedItemToShow = _selectedItemToShow.asStateFlow()
+
 
     init {
         viewModelScope.launch {
@@ -61,18 +65,20 @@ class HomeViewModel @Inject constructor (
                 .mapLatest { cateList ->
                     val itemsInCate = mutableMapOf< Long,StateFlow<List<Item>> >()
                     cateList.forEach {
-                        itemsInCate[it.id] = getItemsOfCategory(it).stateIn(viewModelScope)//(it).first()
+                        itemsInCate[it.id] = getItemsOfCategory(it).stateIn(viewModelScope)
                     }
                     itemsInCate
                 }
                 .stateIn(viewModelScope, SharingStarted.Eagerly, emptyMap<Long,StateFlow<List<Item>>>().toMutableMap() )
         }
-
     }
-
 
     fun setSelectedCategoryID(selectedCategoryID : Long ) {
         savedStateHandle[SELECTED_CATEGORY_KEY] = selectedCategoryID
+    }
+
+    fun setSelectedItemToShow(item : Item) {
+        _selectedItemToShow.value = item
     }
 
     private fun getCategories() =
@@ -104,13 +110,28 @@ class HomeViewModel @Inject constructor (
         actionModeSelectedItemsList.clear()
     }
 
-    fun deleteSelectedCategory() {
-        viewModelScope.launch {
-            selectedCategoryStateFlow.value?.let {
-                itemsStateFlow.value[it.id]?.value?.forEach { item ->
-                    itemsRepository.deleteItem(item)
+    /* @return false when category has items, it is disallowed to delete category that has items. */
+    fun deleteSelectedCategory() : Boolean {
+        val numOfItemsInSelectedCategory = itemsStateFlow.value[selectedCategoryIDStateFlow.value]?.value?.size ?: 0
+        if ( numOfItemsInSelectedCategory > 0 ) {
+            return false
+        } else {
+            viewModelScope.launch {
+                selectedCategoryStateFlow.value?.let {
+                    itemsStateFlow.value[it.id]?.value?.forEach { item ->
+                        itemsRepository.deleteItem(item)
+                    }
+                    itemsRepository.deleteCategory(it)
                 }
-                itemsRepository.deleteCategory(it)
+            }
+            return true
+        }
+    }
+
+    fun saveItemChanges( item:Item ) {
+        selectedCategoryStateFlow.value?.let {
+            viewModelScope.launch {
+                itemsRepository.editItem(category = it, item = item)
             }
         }
     }
@@ -125,6 +146,18 @@ class HomeViewModel @Inject constructor (
         }
     }
 
+    fun getSelectedItemsFilePathAndUnselect() : List<String> {
+        val selectedUris = actionModeSelectedItemsList
+            .stream()
+            .map { item ->
+                item.isSelected = false
+                item.imagePath
+            }
+            .collect(Collectors.toList())
+        actionModeSelectedItemsList.clear()
+        return selectedUris.toList()
+
+    }
 
     fun addImageDataToDB(tempImageUri: Uri) {
         val imageFile = tempImageUri.toFile()
@@ -133,7 +166,6 @@ class HomeViewModel @Inject constructor (
 
             imageFile.takeIf { it.exists() }
                 ?.apply {
-
                     selectedCategoryStateFlow?.value?.let {
                     val item =
                         Item(
@@ -149,7 +181,6 @@ class HomeViewModel @Inject constructor (
                     }
                 }
         }
-
     }
 
     fun importImageToAppFolder(inputStream:InputStream, fileExtension:String?) {
